requirements-docker.txt
boto3==1.9.96
botocore==1.12.96
certifi==2018.10.15
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.9
django-storages==1.7.1
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
Pillow==8.1.0
python-dateutil==2.8.0
pytz==2018.5
s3transfer==0.2.0
whitenoise==4.1.2
django-background-tasks==1.2.5
django-reset-migrations==0.4.0
six<2.0.0
django-utils-six==2.0
django-countries==7.1
searchtweets-v2==1.0.7
searchtweets==1.7.6
django-extensions==3.1.1
cx_freeze==6.5.3
pytest==5.2.0
pytest-cov==2.5.1
pytest-dependency==0.5.1
requests==2.24.0
python-twitter==3.5
djutils==1.0.13
TwitterAPI==2.7.5
django-treenode==0.17.0
django-debug-toolbar==3.2.2
nltk~=3.6.2
PyYAML~=5.3.1
tqdm==4.62.2
ipython-cache==0.2.6
djangorestframework==3.12.4
drf-renderer-xlsx==0.4.1
spacy==3.1.2
sentence-transformers==2.1.0
bertopic==0.9.3
numpy==1.21.2
pandas==1.3.3
django-pandas==0.6.6
gensim==4.1.0
django-filter==21.1
django-annoying==0.10.6
twarc==2.8.0
fasttext==0.9.2
